requirements.txt
flask==2.0.1
transformers
torch
numpy
gunicorn==20.1.0
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pywebview==4.3.3
pyinstaller==6.3.0
